backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
google-generativeai==0.8.3
python-multipart==0.0.20
Pillow==11.1.0
PyPDF2==3.0.1
aiofiles==24.1.0
httpx==0.28.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0
PyMySQL==1.1.1
cryptography==43.0.3
twilio
# AWS SDK + Lambda adapter
boto3>=1.34.0
mangum>=0.17.0
# Background job scheduler (SNS deadline alerts — runs daily at 9 AM IST)
apscheduler>=3.10.4
